+++ src/controllers/GameController.ts
@@ -1,7 +1,7 @@
 import { Engine } from '@babylonjs/core';
 
+import BaseController from '@mmorpg/controllers/base/BaseController';
 import ScenesController from '@mmorpg/controllers/ScenesController';
-import BaseController from '@mmorpg/controllers/BaseController';
 import LoadingScene from '@mmorpg/scenes/virtuals/LoadingScene';
 
 /**
