+++ src/controllers/base/BasePlayerCharacterController.ts
@@ -1,16 +1,15 @@
 import { AbstractMesh, Nullable } from '@babylonjs/core';
 
 import { IBasePlayerCharacterControllerConstructorParams } from '@mmorpg/interfaces/controllers/player/IBasePlayerCharacterController';
-import ICharacterRelated from '@mmorpg/interfaces/common-interfaces/ICharacterRelated';
+import BaseCharacterController from '@mmorpg/controllers/base/BaseCharacterController';
 import PlayerCharacter from '@mmorpg/game-objects/characters/PlayerCharacter';
-import BaseController from '@mmorpg/controllers/BaseController';
 
-abstract class BasePlayerCharacterController extends BaseController implements ICharacterRelated {
+abstract class BasePlayerCharacterController extends BaseCharacterController {
 	protected _characterInstance: PlayerCharacter;
 	protected _characterMesh: Nullable<AbstractMesh>;
 
 	constructor(params: IBasePlayerCharacterControllerConstructorParams) {
-		super();
+		super(params);
 		this._characterInstance = params.characterInstance;
 		this._characterMesh = params.characterInstance.rootNode ?? null;
 	}
