+++ src/interfaces/controllers/player/IPlayerCharacterInventoryController.ts
@@ -1,4 +1,4 @@
-import { IBasePlayerCharacterControllerConstructorParams } from './IBasePlayerCharacterController';
+import { IBasePlayerCharacterControllerConstructorParams } from '@mmorpg/interfaces/controllers/player/IBasePlayerCharacterController';
 
 type IPlayerCharacterInventoryControllerConstructorParams = IBasePlayerCharacterControllerConstructorParams;
